Update variable names and descriptions in Terraform scripts
This commit changes several variable names and descriptions in the Terraform scripts. It updates the naming of the 'azurerm_key_vault_access_policy' resource from 'default' to 'current_user' and modifies some descriptions related to the NGINX ingress controller and resource group. This change should improve code readability and understanding.